factor apart SourceDestSubstep
