+++ protocol-designer/src/components/steplist/index.js
@@ -1,5 +1,5 @@
 // @flow
-import StepItem from './StepItem.js'
+import StepItem from './StepItem'
 import StepList from './StepList'
 
 export {
